+++ lib/src/dtos/apple_client_secret.g.dart
@@ -1,0 +1,17 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'apple_client_secret.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+AppleClientSecret _$AppleClientSecretFromJson(Map<String, dynamic> json) =>
+    AppleClientSecret(
+      secret: json['secret'] as String? ?? "",
+    );
+
+Map<String, dynamic> _$AppleClientSecretToJson(AppleClientSecret instance) =>
+    <String, dynamic>{
+      'secret': instance.secret,
+    };